test(cmd): dedupe redeclared coverage tests
cmd/ had three test functions each declared multiple times across files,
causing go vet to fail with "redeclared in this block":

- TestGetBuildDate, TestGetGitCommit, TestSetupLightweightContext,
  TestSetupLightweightContextPreservesExistingContext existed in both
  cmd/root_test.go (twice — top and bottom) and cmd/root_coverage_test.go.
- TestValidXMLFiles existed in both cmd/shared_flags_xml_test.go and
  cmd/shared_flags_coverage_test.go.

Consolidated to one copy each:
- cmd/root_coverage_test.go deleted; bottom half of cmd/root_test.go
  retains the most thorough versions, plus the GetCommandContext+Config
  assertion from the other copies. Removed a bogus
  `cmdCtx.Config.Engine` assertion that referenced a field that doesn't
  exist on the Config struct.
- cmd/shared_flags_coverage_test.go: dropped its duplicate
  TestValidXMLFiles and unused os import. Kept its other tests
  (TestAddDisplayFlags*, TestDeviceType*, TestBuildEffectiveFormat*).
- cmd/validate_test.go reverted to main — the new copy was identical
  to the existing cmd/validate_coverage_test.go version.

go test ./cmd/... passes.
