+++ cmd/shared_flags_coverage_test.go
@@ -124,6 +124,7 @@ func TestDeviceTypeUnknownReturnsError(t *testing.T) {
 	assert.Contains(t, err.Error(), "unsupported device type")
 }
 
+
 // TestBuildEffectiveFormatCoverage tests the format building logic.
 func TestBuildEffectiveFormatCoverage(t *testing.T) {
 	tests := []struct {
